% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{attachment_data}
\alias{attachment_data}
\title{Get Files Attached to Bug Reports}
\usage{
attachment_data(
  bug_id,
  attachment_id = NULL,
  obsolete = FALSE,
  patch_only = FALSE,
  output_dir = NULL,
  output_file = NULL,
  base_url = "https://bugs.r-project.org/bugzilla3/rest/bug/"
)
}
\arguments{
\item{bug_id}{\acronym{ID} of one bug (numeric or \code{character}).}

\item{attachment_id}{Optional \acronym{ID} of one attachment (numeric or
\code{character}). If given, \code{bug_id} is ignored.}

\item{obsolete}{Return information about attachments marked as obsolete?
A \code{logical} flag. Default is \code{FALSE}, i.e., drop obsolete
attachments. Ignored when \code{attachment_id} is given.}

\item{patch_only}{Only return information about files marked as a patch?
A \code{logical} flag. Default is \code{FALSE}, i.e., include all kinds of
attachments. Ignored when \code{attachment_id} is given.}

\item{output_dir}{file system directory where attachments will be saved
(\code{character} string). If necessary, this function will create one
directory level.}

\item{output_file}{name of file where the attachment will be saved
(\code{character} string). Only takes effect if \code{attachment_id} is
used instead of \code{bug_id}.}

\item{base_url}{\acronym{URL} to Bugzilla \acronym{REST} \acronym{API}
(bug history). Defaults to R Bugzilla.}
}
\value{
If \code{output_dir} or \code{output_file} is specified, the function
  writes to the file system and invisibly returns the \code{output_dir} or
  \code{output_file} parameter (the one that is actually used). Otherwise
  returns a \code{\link{list}} containing the attachments as either
  \code{\link{raw}} vectors or \code{character} strings, depending on the
  file type. The \code{list} has the following vector attributes with one
  item corresponding to each attachment:
  \item{"file_name"}{Name of file (\code{character}).}
  \item{"is_patch"}{Is the file marked as a patch? (\code{logical})}
  \item{"content_type"}{\acronym{MIME} type of file (\code{character}).}
}
\description{
Gets the attachments of a single bug or any single attachment selected by
its \acronym{ID}. By default, the R bug tracker is queried.
}
\examples{
\dontrun{
# Get attachments related to a bug
data17398b <- attachment_data(17398)

# Get a single attachment
data2328a <- attachment_data(attachment_id = 2328)

# Save files to a temporary directory
dir17148 <- attachment_data(17148, output_dir = tempfile(pattern = "dir"))
unlink(dir17148, recursive = TRUE)

# With and without obsolete files (different code paths used)
data16895b2 <- attachment_data(16895, obsolete = TRUE)
data16895b <- attachment_data(16895)
all(data16895b \%in\% data16895b2)
}
}
