% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{attachment_flags}
\alias{attachment_flags}
\title{Get Flags in Attachment Metadata}
\usage{
attachment_flags(
  attachment_id,
  base_url = "https://bugs.r-project.org/bugzilla3/rest/bug/"
)
}
\arguments{
\item{attachment_id}{\acronym{ID} of one attachment (numeric or
\code{character}).}

\item{base_url}{\acronym{URL} to Bugzilla \acronym{REST} \acronym{API}
(bug history). Defaults to R Bugzilla.}
}
\value{
a \code{\link{tibble}}, with rows representing flags
  and the following columns representing their properties:
  \item{id}{\code{ID} of flag (\code{integer}).}
  \item{name}{Name of flag (\code{character}).}
  \item{type_id}{Type \code{ID} (\code{integer}).}
  \item{creation_date}{Date and time of creation (\code{\link{POSIXct}}).}
  \item{modification_date}{Date and time of last modification
    (\code{\link{POSIXct}}).}
  \item{status}{Status of flag (\code{character}).}
  \item{setter}{Who created or last modified the flag? (\code{character})}
  \item{requestee}{The account requested to handle the flag
    (\code{character}). Empty if not set.}
}
\description{
Returns the flags embedded in the metadata related to a single attachment
of a bug report. By default, the R bug tracker is queried.
}
\examples{
\dontrun{
freezilla <- "https://bugs.freebsd.org/bugzilla/rest/bug/"
flags_freebsd <- attachment_flags(183342, base_url = freezilla)
print(flags_freebsd)
}
}
\seealso{
\code{\link{attachment_info}}
}
