% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachments.R
\name{attachment_info}
\alias{attachment_info}
\title{Get Attachment Metadata}
\usage{
attachment_info(bug_id, attachment_id = NULL, obsolete = FALSE,
  patch_only = FALSE,
  base_url = "https://bugs.r-project.org/bugzilla3/rest/bug/")
}
\arguments{
\item{bug_id}{\acronym{ID} of one bug (numeric or \code{character}).}

\item{attachment_id}{Optional \acronym{ID} of one attachment (numeric or
\code{character}). If given, \code{bug_id} is ignored.}

\item{obsolete}{Return information about attachments marked as obsolete?
A \code{logical} flag. Default is \code{FALSE}, i.e., drop obsolete
attachments. Ignored when \code{attachment_id} is given.}

\item{patch_only}{Only return information about files marked as a patch?
A \code{logical} flag. Default is \code{FALSE}, i.e., include all kinds of
attachments. Ignored when \code{attachment_id} is given.}

\item{base_url}{\acronym{URL} to Bugzilla \acronym{REST} \acronym{API}
(bug history). Defaults to R Bugzilla.}
}
\value{
a \code{\link{tibble}}, with rows representing attachments
  and the following columns representing their properties:
  \item{bug_id}{Bug \code{ID} (\code{integer}).}
  \item{id}{Attachment \code{ID} (\code{integer}).}
  \item{summary}{Short description of attachment (\code{character}).}
  \item{file_name}{Filename of attachment (\code{character}).}
  \item{size}{Size of attachment in bytes (\code{integer}).}
  \item{content_type}{\acronym{MIME} type of attachment (\code{character}).}
  \item{creator}{Who created it? (\code{character})}
  \item{creation_time}{Date and time of creation (\code{\link{POSIXct}}).}
  \item{last_change_time}{Date and time of last change (\code{\link{POSIXct}}).}
  \item{is_patch}{Is it a patch? (\code{logical})}
  \item{is_obsolete}{Is it obsolete? (\code{logical})}
  \item{is_private}{Is it private? (\code{logical})}
  \item{any_flags}{Does the attachment have any flags? (\code{logical})}
}
\description{
Returns the metadata related to the attachments of a single bug or any
single attachment selected by its \acronym{ID}. By default, the R bug
tracker is queried.
}
\examples{
\dontrun{
att17398b <- attachment_info(17398)
att2326a <- attachment_info(attachment_id = 2326)
identical(att17398b[att17398b$id == 2326, ], att2326a)
}
}
\seealso{
\code{\link{attachment_flags}}
}
