% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug_history.R
\name{bug_history}
\alias{bug_history}
\title{Get Change History of Bug}
\usage{
bug_history(
  bug_id,
  new_since = NULL,
  base_url = "https://bugs.r-project.org/bugzilla3/rest/bug/"
)
}
\arguments{
\item{bug_id}{\acronym{ID} or alias of one bug (numeric or \code{character}).}

\item{new_since}{Discard changes older than this. A \code{character} string
with date in format \verb{YYYY-MM-DD} or date and time as
\verb{YYYY-MM-DDTHH24:MI:SSZ}. The latter is year-month-date, letter
\code{"T"}, elements of time (hours, minutes and seconds) separated by
\code{:}, and a final letter \code{"Z"} to mark that \acronym{UTC} time is
being used. The standard \link[=DateTimeClasses]{date-time classes} are
also supported. Numeric values are accepted and interpreted as number of
days relative to the current date. The default \code{NULL} means no limit.}

\item{base_url}{\acronym{URL} to Bugzilla \acronym{REST} \acronym{API}
(bug history). Defaults to R Bugzilla.}
}
\value{
a \code{\link{tibble}}, with rows representing historical changes to
  the bug (in chronological order) and the following columns describing the
  details of the change:
  \item{when}{Date and time of change (\code{\link{POSIXct}}).}
  \item{who}{Email address of person who made the change (\code{character}).}
  \item{field_name}{Field that was changed (\code{character}).}
  \item{removed}{Previous value of the field (\code{character}).}
  \item{added}{New value of the field (\code{character}).}
  \item{attachment_id}{\acronym{ID} of attachment that was changed
    (\code{integer}, \code{NA_integer_} if the change was not about an
    attachment).}
}
\description{
Returns the change history of a single bug. Throughout the life of a bug
report, its status may get changed from UNCONFIRMED to NEW or ASSIGNED, and
finally to CLOSED. New people may be added to the \acronym{CC} list of people
subscribed to notifications about the bug. These and other changes are
recorded, and the changes can be inspected with this function. By default,
the R bug tracker is queried.
}
\details{
If several properties (fields) of the bug were changed at once, the result
will show duplicated values in the \code{"when"} and \code{"who"} columns.
}
\examples{
\dontrun{
hist1 <- bug_history(1)
}
}
