% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bug_info.R
\name{bug_info}
\alias{bug_info}
\title{Get Information about Selected Bugs}
\usage{
bug_info(bug_ids, base_url = "https://bugs.r-project.org/bugzilla3/rest/bug/")
}
\arguments{
\item{bug_ids}{One or more numeric (or coercible to numeric) bug
\acronym{ID}s}

\item{base_url}{\acronym{URL} to Bugzilla \acronym{REST} \acronym{API}
(get bug). Defaults to R Bugzilla.}
}
\value{
a \code{\link{tibble}}, with rows representing bugs (in the order
  specified by the \code{bug_ids} argument) and the following columns
  representing their properties:
  \item{id}{Unique bug \code{ID} (\code{integer}).}
  \item{summary}{Short description of bug (\code{character}).}
  \item{creator}{Who created the bug report? (name followed by email in angle
    brackets, \code{character})}
  \item{is_open}{Is the bug open? (\code{logical})}
  \item{product}{Product where the bug occurs (\code{character}).}
  \item{component}{Which part of the product is affected? (\code{character})}
  \item{assignee}{To whom is the bug assigned? (name followed by email in
    angle brackets, \code{character})}
  \item{status}{Status of bug (\code{character}): \code{"UNCONFIRMED"},
    \code{"NEW"}, \code{"CLOSED"}, \ldots}
  \item{is_confirmed}{Has the bug been confirmed? (\code{logical})}
  \item{resolution}{Together with \code{"status"}, defines the current state
    of the bug: \code{"FIXED"}, \code{"WONTFIX"}, \code{"WORKSFORME"},
    \code{"INVALID"}, \ldots}
  \item{created}{Date and time of creation (\code{\link{POSIXct}}).}
  \item{changed}{Date and time of last change (\code{\link{POSIXct}}).}
  \item{alias}{Any aliases the bug may have, separated by \code{";"}
    (\code{character}).}
  \item{version}{Version of product related to bug (\code{character}).}
  \item{os}{Operating system where the bug occurs (\code{character}).}
  \item{hardware}{Type of hardware exhibiting the bug (\code{character}).}
  \item{priority}{Priority of the bug (\code{character}). Typically
    \code{"P1"} means the highest priority and \code{"P5"} the lowest
    priority.}
  \item{severity}{Severity of the bug (\code{character}): \code{"Blocker"},
    \code{"Critical"}, \code{"Major"}, \code{"Normal"}, \code{"Minor"},
    \code{"Trivial"}, \code{"Enhancement"}.}
  \item{url}{\acronym{URL} related to bug (\code{character}, possibly
    empty \code{""}).}
  \item{depends_on}{A list of bugs (numeric \acronym{ID} as \code{character})
    which have to be solved before this one. Uses \code{";"} as separator.
    Possibly empty \code{""}.}
  \item{blocks}{A list of bugs which cannot be solved before this one.
    Formatted like \code{"depends_on"}.}
  \item{cc}{List of interested parties subscribed to updates about the bug
    report. A \code{character} string, with people separated by \code{";"}.
    Each person is shown as a name followed by an email address in angle
    brackets. Possibly empty \code{""}.}
  \item{see_also}{List of \acronym{URL}s pointing to related bug reports.
    Uses the space \code{" "} as separator. Possibly empty \code{""}.}
  \item{dupe_of}{If the bug report is a duplicate of a previous issue, this
    contains the bug \acronym{ID} of the earlier report as a \code{character}
    string. Otherwise an empty string: \code{""}.}
}
\description{
Returns more detailed bug information than \code{\link{list_bugs}}. By
default, the R bug tracker is queried.
}
\examples{
\dontrun{
info1 <- bug_info(1)
}
}
