% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bugs.R
\name{list_bugs}
\alias{list_bugs}
\title{List Bugs in a Bugzilla Database}
\usage{
list_bugs(category = c("open", "closed", "other", "all"), status = NULL,
  changed_from = NULL, changed_to = "Now", product = NULL,
  component = NULL, sorting = c("changed", "id", "importance", "assignee",
  "priority", "severity"), reverse = match.arg(sorting) == "changed",
  base_url = "https://bugs.r-project.org/bugzilla3/buglist.cgi")
}
\arguments{
\item{category}{A \code{character} string: what kind of bugs should be
listed? See \sQuote{Details}.}

\item{status}{An optional \code{character} vector. If set, overrides
\code{category} and defines the exact values for accepted bug status.}

\item{changed_from}{Optional \code{character} string, date-time object of
numeric value to limit the time frame of bugs shown, based on the change
date and time of the bug. This is the start of the search interval.
\code{NULL} (the default) means no limit. See \sQuote{Details}.}

\item{changed_to}{Like \code{changed_from}, but this is the other endpoint.
Defaults to the present (no limit).}

\item{product}{An optional \code{character} string which limits the bug list
to a single product.}

\item{component}{Like \code{product}, but only show a single component.}

\item{sorting}{A \code{character} string stating the desired sorting of the
results. The default \code{"changed"} orders according to the date and time
of last change, while \code{"id"} uses the numerical order of the bug
\acronym{ID}s. Also \code{"priority"} and \code{"severity"} are simple
choices which sort according to one property of each bug. See
\sQuote{Details} for the other choices.}

\item{reverse}{A \code{logical} flag: reverse the order of results? The
default for \code{sorting = "changed"} is \code{TRUE} (most recently
changed first), otherwise \code{FALSE}.}

\item{base_url}{\acronym{URL} of a Bugzilla bug list page. Defaults to R
Bugzilla.}
}
\value{
a \code{\link{tibble}}, with rows representing bugs and the following
  columns representing their properties:
  \item{id}{Unique bug \code{ID} (\code{integer}).}
  \item{summary}{Short description of bug (\code{character}).}
  \item{link}{\code{URL} to detailed bug report web page (\code{character}).}
  \item{product}{Product where the bug occurs (\code{character}).}
  \item{component}{Which part of the product is affected? (\code{character})}
  \item{assignee}{To whom is the bug assigned? (\code{character})}
  \item{status}{Abbreviated status of bug (\code{character}):
    \code{"UNCO"}(NFIRMED), \code{"NEW"}, \code{"CLOS"}(ED), \ldots}
  \item{resolution}{Together with \code{"status"}, defines the current state
    of the bug: \code{"FIXE"}(D), \code{"WONT"}(FIX), \code{"WORK"}(SFORME),
    \code{"INVA"}(LID), \ldots}
  \item{changed}{(Date and) time of last change (\code{character}). The
    format may be irregular.}
}
\description{
Shows the \acronym{ID}s, short descriptions (summaries), and other
information about bugs in a Bugzilla database. The bugs can be filtered by
status and date of last change. By default, the R bug tracker is queried.
}
\details{
The \code{category} option accepts a number of choices. The default
(\code{"open"}) returns bugs that have one of the following as their status:
NEW, ASSIGNED, REOPENED, UNCONFIRMED. The other choices are \code{"closed"}
which covers bugs with the CLOSED status, \code{"other"} for bugs with a
WISH, RESOLVED, or VERIFIED status, and \code{"all"} for listing bugs with
any status.

To limit the results by date and time of last change, use \code{changed_from}
and \code{changed_to}. These options accept a date in the standard
\verb{YYYY-MM-DD} (year, month, day) notation or a relative date-time, for
example \code{"-2d"} for \dQuote{two days ago} including the time of day. The
following abbreviations can be used in relative date-times: \code{"y"} for
year, \code{"m"} for month, \code{"w"} for week, \code{"d"} for day and
\code{"h"} for hour. Numeric values may also be used, and they are
interpreted as days relative to the current date. The standard
\link[=DateTimeClasses]{date-time classes} are also supported, but only the
date part will be utilized. Note that the option to use numeric and date-time
types is a convenience feature provided by this function. Thus, they are
eventually transmitted to the server as \verb{YYYY-MM-DD}, which apparently
refers to the beginning of the day. See \sQuote{Examples}.

The \code{sorting} argument has two choices which sort according to multiple
properties of each bug: \code{"importance"} uses \code{"priority"} and
\code{"severity"} (in this order), whereas with \code{"assignee"} the sort
order is primarily based on the person to whom the bug is assigned to, but
also on the bug status, priority and \acronym{ID} (in this order).

This function extracts information from an \acronym{HTML} document
presenting the bug list. Therefore some results are presented in an
abbreviated form. More detailed information about selected bugs can
be obtained with \code{\link{bug_info}} and
\code{\link{bug_history}}.
}
\examples{
\dontrun{
## Several ways to get open bugs without time limit
bugs1a <- list_bugs() # changed_to = "Now"
bugs1b <- list_bugs(changed_to = 1)
bugs1c <- list_bugs(changed_to = Sys.Date() + 1)
bugs1d <- list_bugs(changed_to = as.character(Sys.Date() + 1))

## List bugs in reverse order.
## Note that reverse = TRUE is default for sorting = "changed", which
## means most recent first. So, reverse = FALSE is oldest first.
bugs2 <- list_bugs(reverse = FALSE)
identical(bugs1a$id, rev(bugs2$id)) # TRUE if bug database did not change

## Get all bugs changed during the last week, sorted according to priority
bugs3 <- list_bugs(category = "all", changed_from = "-1w",
                   sorting = "priority")

## Override predefined bug categories to look for ASSIGNED bugs
bugs4 <- list_bugs(status = "ASSIGNED")
}
}
