% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_changed.R
\name{list_changed}
\alias{list_changed}
\title{List Bugzilla Reports Opened or Closed in a Given Period}
\usage{
list_changed(
  type = c("opened", "opened_confirmed", "closed", "closed_fixed"),
  from = to - 120 - 7 * 24 * 3600,
  to = Sys.time() + 60,
  exclude_components = "Wishlist",
  list_url = NULL,
  info_url = NULL
)
}
\arguments{
\item{type}{Type of changes to list (\code{character}). See \sQuote{Details}.}

\item{from}{Starting point of search interval. Look for changes that occurred
at this moment or after it. Must be \code{character}, numeric, or a value
belonging to one of the \link[=DateTimeClasses]{date-time classes}. See the
description of \code{new_since} in \code{\link{bug_history}} for details.}

\item{to}{Like \code{from}, but end point of time interval.}

\item{exclude_components}{Don't show bugs with the component field matching
one of the given \code{character} strings.}

\item{list_url}{Optional \acronym{URL} to use for the \code{base_url} option
of \code{\link{list_bugs}}. Defaults to using the R Bugzilla.}

\item{info_url}{Optional \acronym{URL} for \code{base_url} in
\code{\link{bug_info}} and \code{\link{bug_history}}. If not explicitly
set, the value of this option will be derived from a user-specified
\code{list_url}. Defaults to using the R Bugzilla.}
}
\value{
A list of bugs matching the criteria. The format is the same as used
  by \code{\link{list_bugs}}.
}
\description{
Specify the type of status changes you are interested in and the limits of a
time interval during which the status change must have occurred. The function
will return a list of bugs matching these criteria. By default, the R bug
tracker is queried.
}
\details{
The \code{type} option specifies what kind of changes will be detected:
\describe{
\item{\code{"opened"}}{Bug report was created during the time frame
  (\code{from} \ldots \code{to}) and is still open.}
\item{\code{"opened_confirmed"}}{Like \code{"opened"}, and the bug is
  confirmed.}
\item{\code{"closed"}}{The last status change of the bug occurred during the
  time frame, and the bug is still CLOSED.}
\item{\code{"closed_fixed"}}{Like \code{"closed"}, and the resolution of the
  bug is FIXED.}
}
}
\examples{
\dontrun{
fixed_jan_2017 <- list_changed("closed_fixed",
                               from = "2017-01-01", to = "2017-02-01")
}
}
